duplicate block content or selection
